requirements.txt
# Deep Learning Frameworks
tensorflow
torch

# Utilities
torchinfo
tensorboard
jupyter-notebook

# Data Management
scikit-learn
matplotlib
pandas
pillow
numpy

# Datasets
kaggle

# Object Detection Frameworks
super-gradients
ultralytics